Wrap markdown-in-html to be rendered in a div before rendering. #8
As noted in the bug, it's necessary to wrap the markdown-in-html in a tag of some kind
so that nested lists render correctly.
This is done automatically if using "paragraph" format (Thunderbird default), but when
using "body text" there's no wrapper. Setting Thunderbird to use paragraph format will
get around the bug as well.