Fix multiple problems with options and migrations...
The short version is that webext-options-sync does not handle async migration functions, so that
was just a disaster.
Implemented async migration functions independent of webext-options-sync/mailext-options-sync.
Then there was a major issue with booleans being stored as strings, so of course
they always evaluated to "true" and messed up the options page.