Disable inline directive syntax. #138
I personally like the inline directive syntax as it allows for things
like colored text :span[red text]{style='color: red'}`. It's not a
standard feature of Markdown. The author of marked-directive references
https://talk.commonmark.org/t/generic-directives-plugins-syntax/444.

The inline directive implementation in marked-directive is lacking
in that the bracketed [content] ends up being optional due to a
incorrect regex.

Inline directive syntax is disabled until I get a chance to fix it up.
::block and :::container syntax are still available.